+++ linux.sh
@@ -3,7 +3,7 @@
 VERSION=$1
 [ -z "$GITHUB_WORKSPACE" ] && GITHUB_WORKSPACE="$( cd "$( dirname "$0" )"/.. && pwd )"
 
-if [ "$VERSION" == "10.6.194" ]; then 
+if [ "$VERSION" == "10.6.194" -o "$VERSION" == "11.8.172" ]; then 
     sudo apt-get install -y \
         pkg-config \
         git \
@@ -33,13 +33,13 @@ fi
 cd ~
 echo "=====[ Getting Depot Tools ]====="	
 git clone -q https://chromium.googlesource.com/chromium/tools/depot_tools.git
-if [ "$VERSION" != "10.6.194" ]; then 
+if [ "$VERSION" != "10.6.194" -a "$VERSION" != "11.8.172" ]; then 
     cd depot_tools
     git reset --hard 8d16d4a
     cd ..
 fi
 export DEPOT_TOOLS_UPDATE=0
-if [ "$VERSION" == "10.6.194" ]; then 
+if [ "$VERSION" == "10.6.194" -o "$VERSION" == "11.8.172" ]; then 
     export PATH=$(pwd)/depot_tools:$PATH
 else
     export PATH=$(pwd)/depot_tools:$(pwd)/depot_tools/.cipd_bin/2.7/bin:$PATH
@@ -68,7 +68,7 @@ node $GITHUB_WORKSPACE/node-script/patchs.js . $VERSION
 
 echo "=====[ Building V8 ]====="
 
-if [ "$VERSION" == "10.6.194" ]; then 
+if [ "$VERSION" == "10.6.194" -o "$VERSION" == "11.8.172" ]; then 
     python ./tools/dev/v8gen.py x64.release -vv -- '
     is_debug = false
     v8_enable_i18n_support= false
